feat: use redux for this
