+++ src/legacy/state/orders/hooks.ts
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo } from 'react'
+import { useCallback, useMemo } from 'react'
 
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 
@@ -16,7 +16,6 @@ import {
   AddOrUpdateOrdersParams,
   addPendingOrder,
   cancelOrdersBatch,
-  CONFIRMED_STATES,
   expireOrdersBatch,
   fulfillOrdersBatch,
   FulfillOrdersBatchParams,
@@ -32,12 +31,11 @@ import {
   setOrderCancellationHash,
   updatePresignGnosisSafeTx,
   UpdatePresignGnosisSafeTxParams,
+  clearOrdersStorage,
 } from './actions'
-import { MAX_ITEMS_PER_STATUS, ORDER_STORAGE_KEY } from './consts'
 import {
   getDefaultNetworkState,
   ORDER_LIST_KEYS,
-  OrderLists,
   OrderObject,
   ORDERS_LIST,
   OrdersState,
@@ -416,64 +414,8 @@ export const useSetIsOrderRefundedBatch = (): SetIsOrderRefundedBatchCallback =>
  * Hook that "cleans" orders local storage
  * Related issue https://github.com/cowprotocol/cowswap/issues/2690
  *
- * It runs when the parent component is unmounted
  */
-
-export const useCleanOrdersStorage = () => {
-  useEffect(() => {
-    _cleanOrdersStorage()
-    return _cleanOrdersStorage
-  }, [])
-}
-
-function _cleanOrdersStorage() {
-  // Get orders state from local storage
-  const ordersCache = localStorage.getItem(ORDER_STORAGE_KEY)
-
-  if (!ordersCache) return
-
-  const parsedState = JSON.parse(ordersCache)
-
-  // Iterate by each chain
-  Object.keys(parsedState).forEach((chainId) => {
-    const orderListByChain: OrderLists = parsedState[chainId]
-
-    // Iterate order statuses we want to clean up
-    CONFIRMED_STATES.forEach((status) => {
-      const orders = orderListByChain[status]
-
-      // Sort by expiration time
-      const ordersCleaned = Object.values(orders)
-        .sort(_orderSorterByExpirationTime)
-        // Take top n orders
-        .slice(0, MAX_ITEMS_PER_STATUS)
-        // Return back to appropriate data structure
-        .reduce<PartialOrdersMap>(_toPartialsOrderMap, {})
-
-      // Update parts of the state, with the "cleaned" ones
-      parsedState[chainId][status] = ordersCleaned
-    })
-  })
-
-  // Stringify data and save to storage
-  localStorage.setItem(ORDER_STORAGE_KEY, JSON.stringify(parsedState))
-}
-
-function _orderSorterByExpirationTime(a: OrderObject | undefined, b: OrderObject | undefined) {
-  const validToA = Number(a?.order.validTo)
-  const validToB = Number(b?.order.validTo)
-
-  if (!validToA || !validToB) {
-    return -1
-  }
-
-  const expirationTimeB = Number(new Date(validToB * 1000))
-  const expirationTimeA = Number(new Date(validToA * 1000))
-
-  return expirationTimeB - expirationTimeA
-}
-
-function _toPartialsOrderMap(acc: PartialOrdersMap, element: OrderObject | undefined) {
-  element && (acc[element.id] = element)
-  return acc
+export const useClearOrdersStorage = () => {
+  const dispatch = useDispatch<AppDispatch>()
+  return useCallback(() => dispatch(clearOrdersStorage()), [dispatch])
 }
