+++ src/legacy/state/orders/reducer.ts
@@ -28,8 +28,10 @@ import {
   updateLastCheckedBlock,
   updateOrder,
   updatePresignGnosisSafeTx,
+  clearOrdersStorage,
+  CONFIRMED_STATES,
 } from './actions'
-import { ContractDeploymentBlocks } from './consts'
+import { ContractDeploymentBlocks, MAX_ITEMS_PER_STATUS } from './consts'
 
 export interface OrderObject {
   id: OrderID
@@ -201,6 +203,25 @@ function cancelOrderInState(state: Required<OrdersState>, chainId: ChainId, orde
   orderObject.order.isCancelling = false
 
   addOrderToState(state, chainId, id, 'cancelled', orderObject.order)
+}
+
+function _orderSorterByExpirationTime(a: OrderObject | undefined, b: OrderObject | undefined) {
+  const validToA = Number(a?.order.validTo)
+  const validToB = Number(b?.order.validTo)
+
+  if (!validToA || !validToB) {
+    return -1
+  }
+
+  const expirationTimeB = Number(new Date(validToB * 1000))
+  const expirationTimeA = Number(new Date(validToA * 1000))
+
+  return expirationTimeB - expirationTimeA
+}
+
+function _toPartialsOrderMap(acc: PartialOrdersMap, element: OrderObject | undefined) {
+  element && (acc[element.id] = element)
+  return acc
 }
 
 const initialState: OrdersState = {}
@@ -482,6 +503,32 @@ export default createReducer(initialState, (builder) =>
 
       ids.forEach((id) => {
         deleteOrderById(state, chainId, id)
+      })
+    })
+    .addCase(clearOrdersStorage, (state) => {
+      Object.keys(state).forEach((_chainId) => {
+        const chainId = _chainId as unknown as ChainId
+        const orderListByChain = state[chainId]
+
+        // Iterate order statuses we want to clean up
+        CONFIRMED_STATES.forEach((status) => {
+          const orders = orderListByChain?.[status]
+
+          if (!orders) {
+            return
+          }
+
+          // Sort by expiration time
+          const ordersCleaned = Object.values(orders)
+            .sort(_orderSorterByExpirationTime)
+            // Take top n orders
+            .slice(0, MAX_ITEMS_PER_STATUS)
+            // Return back to appropriate data structure
+            .reduce<PartialOrdersMap>(_toPartialsOrderMap, {})
+
+          // Update parts of the state, with the "cleaned" ones
+          orderListByChain[status] = ordersCleaned
+        })
       })
     })
 )
