+++ src/legacy/state/orders/updaters/GpOrdersUpdater.ts
@@ -8,7 +8,7 @@ import { GP_ORDER_UPDATE_INTERVAL, NATIVE_CURRENCY_BUY_ADDRESS, NATIVE_CURRENCY_
 import { useAllTokens } from 'legacy/hooks/Tokens'
 import { useTokenLazy } from 'legacy/hooks/useTokenLazy'
 import { Order, OrderStatus } from 'legacy/state/orders/actions'
-import { useAddOrUpdateOrders, useCleanOrdersStorage } from 'legacy/state/orders/hooks'
+import { useAddOrUpdateOrders, useClearOrdersStorage } from 'legacy/state/orders/hooks'
 import { computeOrderSummary } from 'legacy/state/orders/updaters/utils'
 import { classifyOrder, OrderTransitionStatus } from 'legacy/state/orders/utils'
 
@@ -202,7 +202,7 @@ function _filterOrders(
  * - Persist the new tokens and orders on redux
  */
 export function GpOrdersUpdater(): null {
-  useCleanOrdersStorage()
+  const clearOrderStorage = useClearOrdersStorage()
 
   const { account, chainId } = useWalletInfo()
   const allTokens = useAllTokens()
@@ -263,6 +263,14 @@ export function GpOrdersUpdater(): null {
       updateOrders(chainId, account)
     }
   }, [account, chainId, tokensAreLoaded, updateOrders])
+
+  useEffect(() => {
+    clearOrderStorage()
+
+    return function () {
+      clearOrderStorage()
+    }
+  }, [clearOrderStorage])
 
   return null
 }
